+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+
+## [0.2.0] - 2024-06-30
+### Added
+- ScriptElement class
+- Page.Head.loadScript
+- Page.Body.loadScript
+- Script attributes like ID
+- Script prepend/append options
+
+### Changed
+- Renamed IRouteHandler to IModule
+
+
+
 ## [0.1.1] - 2024-06-29
 ### Added
 - Global SSE lib path